worker/requirements.txt
redis==5.2.1
rq==2.1.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
openai==1.63.2
requests==2.32.3
boto3==1.36.13
pydantic-settings==2.7.1
python-dotenv==1.0.1
pillow==11.1.0
opencv-python-headless==4.10.0.84